reviews almost done

+++ app/Http/Controllers/CartWishlistController.php
@@ -11,6 +11,7 @@ use App\DeliveryType;
 
 use App\Product;
 use App\Purchase;
+use Session;
 
 
 class CartWishlistController extends Controller{
@@ -19,14 +20,40 @@ class CartWishlistController extends Controller{
         
         try {
             if(Auth::check()) {
+                
                 $user = Auth::user();
                 $this->authorize('edit', $user);
+                if(session('wishlist', NULL)) {
+                    $products = $user->wishlist()->pluck('id');
+                    $products_id = session('wishlist');
+                    foreach($products_id as $id) {
+                        foreach($products as $hasid) {
+                            $id = (int) $id;
+                            if($id != $hasid) {
+                                $product = Product::findOrFail($id);
+                                $user->wishlist()->attach($product);
+                                break;
+                            }
+                        }
+                    }
+                    session()->forget('wishlist');
+                    
+                }
                 $products = $user->wishlist()->get();
-                return view('pages.wishlist', ['products' => $products]);
-
+                
             } else {
-                $request->cookie('name');
+
+                $products = array();
+                if(session('wishlist', NULL)){
+                    $products_id = session('wishlist');
+                    foreach($products_id as $id) {
+                        $product = Product::findOrFail($id);
+                        array_push($products, $product);
+                    }
+                }
+
             }
+            return view('pages.wishlist', ['products' => $products]);
         }catch(\Exception $e) {
             return response(json_encode("Error showing wishlist"), 500);
         }
@@ -48,13 +75,24 @@ class CartWishlistController extends Controller{
                 $user = Auth::user();
                 $this->authorize('edit', $user);
                 if($user->wishlist()->where('product_id', $request->id)->count() > 0){
-                    return response(json_encode("You already have this product in your wishlist"), 500);
+                    return response(json_encode("You already have this product in your wishlist"), 401);
                 }
                 $user->wishlist()->attach($product);
 
             } else {
-
-
+                if(session('wishlist', NULL)) {
+                    $cookies = session('wishlist');
+                    if(in_array($request->id, $cookies)){
+                        return response(json_encode("You already have this product in your wishlist"), 401);
+                    }
+                    array_push($cookies, $request->id);
+                    session(['wishlist' => $cookies]);
+                }
+                else {
+                    $cookies = array($request->id);
+                    session(['wishlist' => $cookies]);
+                }
+                
             }
             return response(json_encode("Product added to Wishlist"), 200);
         }
@@ -66,38 +104,51 @@ class CartWishlistController extends Controller{
 
 
     public function removeWishlistProduct(Request $request) {
-        try {
-            $user = Auth::user();
-            $this->authorize('edit', $user);
-            $product = $user->wishlist()->where('product_id', $request->id)->first();
+        if(Auth::check()) {
+        
+            try {
+                $user = Auth::user();
+                $this->authorize('edit', $user);
+                $product = $user->wishlist()->where('product_id', $request->id)->first();
 
-        }catch(\Exception $e) {
-            return response(json_encode("Error removing product from wishlist"), 500);
-        }
+            }catch(\Exception $e) {
+                return response(json_encode("Error removing product from wishlist"), 500);
+            }
 
-        if($product != null){
-            $user->wishlist()->detach([$request->id]);
-            return response(json_encode("Product deleted from wishlist"), 200);
-        }
-        else{
-            return response(json_encode("That produt is not on the user's wishlist"), 404);
+            if($product != null){
+                $user->wishlist()->detach([$request->id]);
+                return response(json_encode("Product deleted from wishlist"), 200);
+            }
+            else{
+                return response(json_encode("That produt is not on the user's wishlist"), 404);
+            }
+        } else {
+            if(($products = session('wishlist', NULL)) != false) {
+                if(in_array($request->id, $products)) {
+                    $products = array_diff($products, [$request->id]);
+                }
+                session(['wishlist' => $products]);
+                return response(json_encode("Product deleted from wishlist"), 200);
+            }
         }
     }
 
     public function showCart(){
         
-        try {
-            $user = Auth::user();
-            $this->authorize('edit', $user);
-            $products = $user->cart()->get();
-            $total = $products->sum(function($t){ 
-                return $t->price*$t->pivot->quantity; 
-            });
-            $addresses = $user->addresses()->get();
-            $delivery_types = DeliveryType::get();
-            return view('pages.cart', ['total'=> $total, 'products' => $products,'addresses' => $addresses, 'delivery_types'=> $delivery_types]);
-        }catch(\Exception $e) {
-            return response(json_encode("Error showing cart"), 500);
+        if(Auth::check()) {
+            try {
+                $user = Auth::user();
+                $this->authorize('edit', $user);
+                $products = $user->cart()->get();
+                $total = $products->sum(function($t){ 
+                    return $t->price*$t->pivot->quantity; 
+                });
+                $addresses = $user->addresses()->get();
+                $delivery_types = DeliveryType::get();
+                return view('pages.cart', ['total'=> $total, 'products' => $products,'addresses' => $addresses, 'delivery_types'=> $delivery_types]);
+            }catch(\Exception $e) {
+                return response(json_encode("Error showing cart"), 500);
+            }
         }
         
     }
@@ -219,11 +270,6 @@ class CartWishlistController extends Controller{
             return response(json_encode("Something went wrong checking out"), 500);
         }
     }
-
-    public function __construct() {
-        $this->middleware('auth');
-    }
-
 }
 
 ?>
